Refactored ImprovedInstanceScript into Game out of Instance_Dire_Maul.h
This was being hacked together in the dire maul header file. Which was
silly, but I was being pragmatic.
